Remove Priority menu blinking
The menu is rendered minimised in the server side since the viewport
size is not known, and then expanded in the client side when the space
to render items is known. However, this results in first having the
minimized menu visible and then swapping the expanded menu to replace
it, which causes a frustrating blink on every page load.

The fix hides the menu when JavaScript is enabled (thus having the menu
working when JS is disabled etc.) until the menu is expanded.